requirements.txt
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
folium==0.18.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
xyzservices==2024.9.0
